adds team athletes grid view
