+++ src/components/fantasy-league/team-modal/FantasyLeagueMemberModal.tsx
@@ -13,6 +13,7 @@ import { formatPosition } from "../../../utils/athleteUtils"
 import SecondaryText from "../../shared/SecondaryText"
 import RoundedCard from "../../shared/RoundedCard"
 import NoContentCard from "../../shared/NoContentMessage"
+import { TeamAthletesGridView } from "./views/TeamAthletesGridView"
 
 type Props = {
     isOpen?: boolean,
@@ -152,59 +153,3 @@ export default function FantasyLeagueMemberModal({ onClose, isOpen, member }: Pr
     )
 }
 
-type TeamAthletesViewProps = {
-    roundTeam: FantasyLeagueTeamWithAthletes
-}
-
-function TeamAthletesGridView({ roundTeam }: TeamAthletesViewProps) {
-    return (
-        <div className="bg-green-600 overflow-clip min-h-[920px] max-h-[920px] relative rounded-xl" >
-            <RugbyPitch count={6} />
-
-            <div className="flex flex-row  min-h-[920px] max-h-[920px] items-center justify-center flex-wrap absolute top-0 left-0" >
-
-                {roundTeam?.athletes?.map((a) => {
-                    return (
-                        <div className="flex flex-col gap-2 items-center justify-center" >
-                            <PlayerGameCard
-                                player={a.athlete}
-                            />
-
-                            <p className="font-bold text-white" >{Math.floor(a.score ?? 0)}</p>
-                        </div>
-                    )
-                })}
-            </div>
-        </div>
-    )
-}
-
-
-function TeamAthletesListView({ roundTeam }: TeamAthletesViewProps) {
-    return (
-        <div className="overflow-clip relative rounded-xl gap-2" >
-
-            <div className="flex flex-col gap-2" >
-
-                {roundTeam?.athletes.map((a) => {
-                    return (
-                        <div className="flex flex-row gap-2 items-center justify-between" >
-                            <div className="flex flex-row items-center gap-2" >
-                                <PlayerMugshot
-                                    url={a.athlete.image_url}
-                                />
-
-                                <div>
-                                    <p className="font-bold truncate text-xs" >{a.athlete.player_name}</p>
-                                    <SecondaryText className="text-xs" >{a.athlete.position_class ? formatPosition(a.athlete.position_class) : '-'}</SecondaryText>
-                                </div>
-                            </div>
-
-                            <p className="font-bold" >{Math.floor(a.score ?? 0)}</p>
-                        </div>
-                    )
-                })}
-            </div>
-        </div>
-    )
-}
